requirements.txt
google-cloud-storage==2.16.0
pickle5==0.0.12
joblib==1.4.2
numpy==1.26.4
fastapi==0.111.0
uvicorn==0.30.1
scikit-learn==1.2.2
gunicorn==20.1.0
matplotlib
opencv-python-headless==4.10.0.82
skia-python==87.6
Mesa==2.3.0
PyOpenGL==3.1.7
pillow==10.3.0
git-lfs==1.6
requests==2.32.3

